pyproject.toml
[tool.poetry]
name = "api-olx"
version = "0.1.0"
description = ""
authors = ["xolmomin <xolmomin@gmail.com>"]
readme = "README.md"
packages = [{ include = "root" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-yasg = "^1.21.5"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
django-mptt = "^0.14.0"
django-ckeditor = "^6.5.1"
django-modeltranslation = "^0.18.10"
django-debug-toolbar = "^4.1.0"
pillow = "^9.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# -- Example FILE: pyproject.toml
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "root.settings"
# -- recommended but optional:
python_files = "tests.py, test_*.py, *_tests.py"
